+++ src/plone/app/imagecropping/browser/scalingdx.py
@@ -24,6 +24,7 @@ class ImageScalingDX(ScalingOverrides, NFImageScaling):
 
     def scale(self, fieldname=None, scale=None, height=None, width=None,
               direction='thumbnail', **parameters):
+        self._scale_name = scale
         self._need_rescale(fieldname, scale)
         # if direction is 'down' and we have a cropped scale
         # deliver it instead of standard 'down' scale
@@ -31,3 +32,15 @@ class ImageScalingDX(ScalingOverrides, NFImageScaling):
             direction = 'thumbnail'
         return super(ImageScalingDX, self).scale(
             fieldname, scale, height, width, direction, **parameters)
+
+    def _wrap_image(self, data, fmt='PNG', fieldname=None):
+        if not fieldname:
+            return
+        orig_value = getattr(self.context, fieldname)
+        if orig_value is None:
+            return
+        mimetype = 'image/%s' % fmt.lower()
+        value = orig_value.__class__(
+            data, contentType=mimetype, filename=orig_value.filename)
+        value.fieldname = fieldname
+        return value